+++ packages/core-p2p/src/peer.ts
@@ -6,6 +6,7 @@ import delay from "delay";
 import socketCluster from "socketcluster-client";
 import util from "util";
 import { config as localConfig } from "./config";
+import { PeerVerifier } from "./peer-verifier";
 
 export class Peer implements P2P.IPeer {
     public downloadSize: any;
@@ -172,17 +173,7 @@ export class Peer implements P2P.IPeer {
      */
     public async downloadBlocks(fromBlockHeight) {
         try {
-            const response: any = await this.emit("p2p.peer.getBlocks", {
-                params: { lastBlockHeight: fromBlockHeight },
-                headers: this.headers,
-                timeout: 10000,
-            });
-
-            /*const response = await axios.get(`${this.url}/peer/blocks`, {
-                params: { lastBlockHeight: fromBlockHeight },
-                headers: this.headers,
-                timeout: 10000,
-            });*/
+            const response = await this.getPeerBlocks(fromBlockHeight);
 
             this.__parseHeaders(response);
 
@@ -210,24 +201,48 @@ export class Peer implements P2P.IPeer {
     /**
      * Perform ping request on this peer if it has not been
      * recently pinged.
-     * @param  {Number} [delay=5000]
+     * @param  {Number} timeoutMsec operation timeout, in milliseconds
      * @param  {Boolean} force
      * @return {Object}
      * @throws {Error} If fail to get peer status.
      */
-    public async ping(maxDelay, force = false) {
+    public async ping(timeoutMsec, force = false) {
+        const deadline = new Date().getTime() + timeoutMsec;
         if (this.recentlyPinged() && !force) {
             return;
         }
 
-        // TODO use parameter maxDelay
+        // TODO use parameter timeoutMsec
         const body: any = await this.emit("p2p.peer.getStatus", null);
         /*
-        const body = await this.__get("/peer/status", delay || localConfig.get("globalTimeout"));
+        const body = await this.__get("/peer/status", delay);
         */
 
         if (!body) {
             throw new Error(`Peer ${this.ip} is unresponsive`);
+        }
+
+        if (!body.success) {
+            throw new Error(
+                `Erroneous response from peer ${this.ip} when trying to retrieve its status: ` + JSON.stringify(body),
+            );
+        }
+
+        if (process.env.CORE_SKIP_PEER_STATE_VERIFICATION !== "true") {
+            const peerVerifier = new PeerVerifier(this);
+
+            if (deadline <= new Date().getTime()) {
+                throw new Error(
+                    `When pinging peer ${this.ip}: ping timeout (${delay} ms) elapsed ` +
+                        `even before starting peer verification`,
+                );
+            }
+
+            if (!(await peerVerifier.checkState(body, deadline))) {
+                throw new Error(
+                    `Peer state verification failed for peer ${this.ip}, claimed state: ` + JSON.stringify(body),
+                );
+            }
         }
 
         this.lastPinged = dayjs();
@@ -265,9 +280,10 @@ export class Peer implements P2P.IPeer {
     /**
      * Check if peer has common blocks.
      * @param  {[]String} ids
+     * @param {Number} timeoutMsec timeout for the operation, in milliseconds
      * @return {Boolean}
      */
-    public async hasCommonBlocks(ids) {
+    public async hasCommonBlocks(ids, timeoutMsec?: number) {
         try {
             /*let url = `/peer/blocks/common?ids=${ids.join(",")}`;
             if (ids.length === 1) {
@@ -275,13 +291,15 @@ export class Peer implements P2P.IPeer {
             }*/
 
             const body: any = await this.emit("p2p.peer.getCommonBlocks", { ids });
+            // TODO handle timeout see below old implementation
             /*
-            const body = await this.__get(url);
+            const body = await this.__get(url, timeoutMsec);
             */
 
             return body && body.success && !!body.common;
         } catch (error) {
-            this.logger.error(`Could not determine common blocks with ${this.ip}: ${error}`);
+            const sfx = timeoutMsec !== undefined ? ` within ${timeoutMsec} ms` : "";
+            this.logger.error(`Could not determine common blocks with ${this.ip}${sfx}: ${error}`);
         }
 
         return false;
@@ -362,6 +380,21 @@ export class Peer implements P2P.IPeer {
         this.status = response.status;
 
         return response;
+    }
+
+    /**
+     * GET /peer/blocks and return the raw response.
+     * The API is such that the response is supposed to contain blocks at height
+     * afterBlockHeight + 1, afterBlockHeight + 2, and so on up to some limit determined by the peer.
+     * @param  {Number} afterBlockHeight
+     * @return {(Object[]|undefined)}
+     */
+    public async getPeerBlocks(afterBlockHeight: number): Promise<any> {
+        return this.emit("p2p.peer.getBlocks", {
+            params: { lastBlockHeight: afterBlockHeight },
+            headers: this.headers,
+            timeout: 10000,
+        });
     }
 
     private async emit(event, data) {