+++ packages/core-p2p/src/server/versions/internal/handlers/transactions.ts
@@ -9,7 +9,7 @@ const { Transaction } = models;
 /**
  * @type {Object}
  */
-export const verify = {
+export const verify: object = {
     /**
      * @param  {Hapi.Request} request
      * @param  {Hapi.Toolkit} h