+++ packages/core-test-utils/src/config/testnet/plugins.js
@@ -18,7 +18,7 @@ module.exports = {
         connection: {
             host: process.env.CORE_DB_HOST || "localhost",
             port: process.env.CORE_DB_PORT || 5432,
-            database: process.env.CORE_DB_DATABASE || `${process.env.CORE_TOKEN}_development`,
+            database: process.env.CORE_DB_DATABASE || `${process.env.CORE_TOKEN}_unitnet`,
             user: process.env.CORE_DB_USERNAME || process.env.CORE_TOKEN,
             password: process.env.CORE_DB_PASSWORD || "password",
         },
@@ -38,7 +38,7 @@ module.exports = {
     "@arkecosystem/core-p2p": {
         host: process.env.CORE_P2P_HOST || "0.0.0.0",
         port: process.env.CORE_P2P_PORT || 4000,
-        minimumVersion: ">=2.0.0",
+        minimumVersions: [">=2.0.0"],
         minimumNetworkReach: 5,
         coldStart: 5,
     },